+++ src/main/java/tkcy/simpleaddon/common/ClientProxy.java
@@ -5,7 +5,6 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.relauncher.Side;
 
-import tkcy.simpleaddon.api.render.GCYMTextures;
 import tkcy.simpleaddon.common.block.TKCYSAMetaBlocks;
 
 @Mod.EventBusSubscriber(Side.CLIENT)
@@ -14,7 +13,6 @@ public class ClientProxy extends CommonProxy {
     @Override
     public void preLoad() {
         super.preLoad();
-        GCYMTextures.preInit();
     }
 
     @SubscribeEvent